+++ server/router/dairy.js
@@ -10,8 +10,8 @@ connectDB.open(db);
 router.get('/',(req,res) => {
   const selectQuery = "select * from dairy"
   db.query(selectQuery, (err, result) => {
-    res.send(result);
-    // console.log('result', result);
+    if(err) console.log("err",err);
+    else {res.send(result)}
   })
 });
 //==============================================
@@ -22,8 +22,8 @@ router.post('/insert',(req,res) => {
   const image = req.body.image ? req.body.image : 'NULL';
   const insertQuery = `INSERT INTO dairy (user_no, title, content, image) VALUES (1, '${title}', '${content}', '${image}');`
   db.query(insertQuery, (err, result) => {
-    res.send(result);
-    // console.log('result', result);
+    if(err) console.log("err",err);
+    else {res.send(result)}
   })
 });
 //==============================================
@@ -33,8 +33,8 @@ router.post('/delete', (req,res) => {
   console.log("id",id)
   const deleteQuery = `DELETE FROM dairy WHERE dairy_no=${id}`;
   db.query(deleteQuery, (err, result) => {
-    // console.log("deleteQuery",deleteQuery);
-    res.send("삭제완료")
+    if(err) console.log("err",err);
+    else {res.send("삭제완료")}
   })
 });
 //==============================================
